pyproject.toml
[project]
name = "faf-shim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
faf-shim = "main:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
